+++ server/runtime_lua.go
@@ -31,7 +31,6 @@ import (
 	"github.com/gofrs/uuid"
 	"github.com/heroiclabs/nakama-common/api"
 	"github.com/heroiclabs/nakama-common/rtapi"
-	"github.com/heroiclabs/nakama-common/runtime"
 	lua "github.com/heroiclabs/nakama/v3/internal/gopher-lua"
 	"github.com/heroiclabs/nakama/v3/social"
 	"go.uber.org/atomic"
@@ -1018,7 +1017,7 @@ func NewRuntimeProviderLua(logger, startupLogger *zap.Logger, db *sql.DB, protoj
 				return runtimeProviderLua.TournamentReset(ctx, tournament, end, reset)
 			}
 		case RuntimeExecutionModeLeaderboardReset:
-			leaderboardResetFunction = func(ctx context.Context, leaderboard runtime.Leaderboard, reset int64) error {
+			leaderboardResetFunction = func(ctx context.Context, leaderboard *api.Leaderboard, reset int64) error {
 				return runtimeProviderLua.LeaderboardReset(ctx, leaderboard, reset)
 			}
 		}
@@ -1615,17 +1614,13 @@ func (rp *RuntimeProviderLua) TournamentEnd(ctx context.Context, tournament *api
 
 	luaCtx := NewRuntimeLuaContext(r.vm, r.node, r.luaEnv, RuntimeExecutionModeTournamentEnd, nil, 0, "", "", nil, "", "", "")
 
-	tournamentTable := r.vm.CreateTable(0, 17)
+	tournamentTable := r.vm.CreateTable(0, 18)
 
 	tournamentTable.RawSetString("id", lua.LString(tournament.Id))
 	tournamentTable.RawSetString("title", lua.LString(tournament.Title))
 	tournamentTable.RawSetString("description", lua.LString(tournament.Description))
 	tournamentTable.RawSetString("category", lua.LNumber(tournament.Category))
-	if tournament.SortOrder == LeaderboardSortOrderAscending {
-		tournamentTable.RawSetString("sort_order", lua.LString("asc"))
-	} else {
-		tournamentTable.RawSetString("sort_order", lua.LString("desc"))
-	}
+	tournamentTable.RawSetString("sort_order", lua.LString(tournament.SortOrder))
 	tournamentTable.RawSetString("size", lua.LNumber(tournament.Size))
 	tournamentTable.RawSetString("max_size", lua.LNumber(tournament.MaxSize))
 	tournamentTable.RawSetString("max_num_score", lua.LNumber(tournament.MaxNumScore))
@@ -1633,7 +1628,17 @@ func (rp *RuntimeProviderLua) TournamentEnd(ctx context.Context, tournament *api
 	tournamentTable.RawSetString("start_active", lua.LNumber(tournament.StartActive))
 	tournamentTable.RawSetString("end_active", lua.LNumber(tournament.EndActive))
 	tournamentTable.RawSetString("can_enter", lua.LBool(tournament.CanEnter))
-	tournamentTable.RawSetString("next_reset", lua.LNumber(tournament.NextReset))
+	if tournament.NextReset != 0 {
+		tournamentTable.RawSetString("next_reset", lua.LNumber(tournament.NextReset))
+	} else {
+		tournamentTable.RawSetString("next_reset", lua.LNil)
+	}
+	if tournament.PrevReset != 0 {
+		tournamentTable.RawSetString("prev_reset", lua.LNumber(tournament.PrevReset))
+	} else {
+		tournamentTable.RawSetString("prev_reset", lua.LNil)
+	}
+
 	metadataMap := make(map[string]interface{})
 	err = json.Unmarshal([]byte(tournament.Metadata), &metadataMap)
 	if err != nil {
@@ -1649,6 +1654,7 @@ func (rp *RuntimeProviderLua) TournamentEnd(ctx context.Context, tournament *api
 	} else {
 		tournamentTable.RawSetString("end_time", lua.LNumber(tournament.EndTime.Seconds))
 	}
+	tournamentTable.RawSetString("operator", lua.LString(strings.ToLower(tournament.Operator.String())))
 
 	// Set context value used for logging
 	vmCtx := context.WithValue(ctx, ctxLoggerFields{}, map[string]string{"mode": RuntimeExecutionModeTournamentEnd.String()})
@@ -1681,24 +1687,29 @@ func (rp *RuntimeProviderLua) TournamentReset(ctx context.Context, tournament *a
 
 	luaCtx := NewRuntimeLuaContext(r.vm, r.node, r.luaEnv, RuntimeExecutionModeTournamentReset, nil, 0, "", "", nil, "", "", "")
 
-	tournamentTable := r.vm.CreateTable(0, 16)
+	tournamentTable := r.vm.CreateTable(0, 18)
 
 	tournamentTable.RawSetString("id", lua.LString(tournament.Id))
 	tournamentTable.RawSetString("title", lua.LString(tournament.Title))
 	tournamentTable.RawSetString("description", lua.LString(tournament.Description))
 	tournamentTable.RawSetString("category", lua.LNumber(tournament.Category))
-	if tournament.SortOrder == LeaderboardSortOrderAscending {
-		tournamentTable.RawSetString("sort_order", lua.LString("asc"))
-	} else {
-		tournamentTable.RawSetString("sort_order", lua.LString("desc"))
-	}
+	tournamentTable.RawSetString("sort_order", lua.LString(tournament.SortOrder))
 	tournamentTable.RawSetString("size", lua.LNumber(tournament.Size))
 	tournamentTable.RawSetString("max_size", lua.LNumber(tournament.MaxSize))
 	tournamentTable.RawSetString("max_num_score", lua.LNumber(tournament.MaxNumScore))
 	tournamentTable.RawSetString("duration", lua.LNumber(tournament.Duration))
 	tournamentTable.RawSetString("end_active", lua.LNumber(tournament.EndActive))
 	tournamentTable.RawSetString("can_enter", lua.LBool(tournament.CanEnter))
-	tournamentTable.RawSetString("next_reset", lua.LNumber(tournament.NextReset))
+	if tournament.NextReset != 0 {
+		tournamentTable.RawSetString("next_reset", lua.LNumber(tournament.NextReset))
+	} else {
+		tournamentTable.RawSetString("next_reset", lua.LNil)
+	}
+	if tournament.PrevReset != 0 {
+		tournamentTable.RawSetString("prev_reset", lua.LNumber(tournament.PrevReset))
+	} else {
+		tournamentTable.RawSetString("prev_reset", lua.LNil)
+	}
 	metadataMap := make(map[string]interface{})
 	err = json.Unmarshal([]byte(tournament.Metadata), &metadataMap)
 	if err != nil {
@@ -1714,6 +1725,7 @@ func (rp *RuntimeProviderLua) TournamentReset(ctx context.Context, tournament *a
 	} else {
 		tournamentTable.RawSetString("end_time", lua.LNumber(tournament.EndTime.Seconds))
 	}
+	tournamentTable.RawSetString("operator", lua.LString(strings.ToLower(tournament.Operator.String())))
 
 	// Set context value used for logging
 	vmCtx := context.WithValue(ctx, ctxLoggerFields{}, map[string]string{"mode": RuntimeExecutionModeTournamentReset.String()})
@@ -1733,7 +1745,7 @@ func (rp *RuntimeProviderLua) TournamentReset(ctx context.Context, tournament *a
 	return errors.New("Unexpected return type from runtime Tournament Reset hook, must be nil.")
 }
 
-func (rp *RuntimeProviderLua) LeaderboardReset(ctx context.Context, leaderboard runtime.Leaderboard, reset int64) error {
+func (rp *RuntimeProviderLua) LeaderboardReset(ctx context.Context, leaderboard *api.Leaderboard, reset int64) error {
 	r, err := rp.Get(ctx)
 	if err != nil {
 		return err
@@ -1746,16 +1758,27 @@ func (rp *RuntimeProviderLua) LeaderboardReset(ctx context.Context, leaderboard 
 
 	luaCtx := NewRuntimeLuaContext(r.vm, r.node, r.luaEnv, RuntimeExecutionModeLeaderboardReset, nil, 0, "", "", nil, "", "", "")
 
-	leaderboardTable := r.vm.CreateTable(0, 13)
+	leaderboardTable := r.vm.CreateTable(0, 8)
 
-	leaderboardTable.RawSetString("id", lua.LString(leaderboard.GetId()))
-	leaderboardTable.RawSetString("authoritative", lua.LBool(leaderboard.GetAuthoritative()))
-	leaderboardTable.RawSetString("sort_order", lua.LString(leaderboard.GetSortOrder()))
-	leaderboardTable.RawSetString("operator", lua.LString(leaderboard.GetOperator()))
-	leaderboardTable.RawSetString("reset", lua.LString(leaderboard.GetReset()))
-	metadataTable := RuntimeLuaConvertMap(r.vm, leaderboard.GetMetadata())
+	leaderboardTable.RawSetString("id", lua.LString(leaderboard.Id))
+	leaderboardTable.RawSetString("authoritative", lua.LBool(leaderboard.Authoritative))
+	leaderboardTable.RawSetString("sort_order", lua.LString(leaderboard.SortOrder))
+	leaderboardTable.RawSetString("operator", lua.LString(strings.ToLower(leaderboard.Operator.String())))
+	if leaderboard.PrevReset != 0 {
+		leaderboardTable.RawSetString("prev_reset", lua.LString(leaderboard.PrevReset))
+	}
+	if leaderboard.NextReset != 0 {
+		leaderboardTable.RawSetString("next_reset", lua.LString(leaderboard.NextReset))
+	}
+	metadataMap := make(map[string]interface{})
+	err = json.Unmarshal([]byte(leaderboard.Metadata), &metadataMap)
+	if err != nil {
+		rp.Put(r)
+		return fmt.Errorf("failed to convert metadata to json: %s", err.Error())
+	}
+	metadataTable := RuntimeLuaConvertMap(r.vm, metadataMap)
 	leaderboardTable.RawSetString("metadata", metadataTable)
-	leaderboardTable.RawSetString("create_time", lua.LNumber(leaderboard.GetCreateTime()))
+	leaderboardTable.RawSetString("create_time", lua.LNumber(leaderboard.CreateTime.Seconds))
 
 	// Set context value used for logging
 	vmCtx := context.WithValue(ctx, ctxLoggerFields{}, map[string]string{"mode": RuntimeExecutionModeLeaderboardReset.String()})